Bring JSON load helper into main module
